+++ test/config/null_manifest_test.rb
@@ -1,9 +1,9 @@
 require 'test_helper'
 require 'pp'
 
-describe Hanami::Assets::Config::NullDigestManifest do
+describe Hanami::Assets::Config::NullManifest do
   let(:configuration) { Hanami::Assets::Configuration.new }
-  let(:manifest)      { Hanami::Assets::Config::NullDigestManifest.new(configuration) }
+  let(:manifest)      { Hanami::Assets::Config::NullManifest.new(configuration) }
 
   it 'is pretty printable' do
     pp manifest